build: move event reaction classes and files on a per se package
