+++ application/src/test/scala/io/github/positionpal/location/application/reactions/TrackingEventReactionsTest.scala
@@ -1,4 +1,4 @@
-package io.github.positionpal.location.application.geo
+package io.github.positionpal.location.application.reactions
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
@@ -6,11 +6,12 @@ import org.scalatest.matchers.should.Matchers
 class TrackingEventReactionsTest extends AnyFunSpec with Matchers:
 
   import TrackingEventReaction.*
-  import java.util.Date
   import cats.effect.IO
   import cats.effect.unsafe.implicits.global
   import io.github.positionpal.location.domain.DrivingEvents.TrackingEvent
   import io.github.positionpal.location.domain.{DrivingEvents, GPSLocation, Route, UserId}
+
+  import java.util.Date
 
   private val route = Route(DrivingEvents.StartRoutingEvent(Date(), UserId("test"), GPSLocation(0.0, 0.0)))
   private val event: TrackingEvent = TrackingEvent(Date(), UserId("test"), GPSLocation(0.1, 0.1))
